+++ src/Libraries/Nop.Core/Domain/Catalog/ProductWarehouseInventory.cs
@@ -18,11 +18,11 @@
         /// <summary>
         /// Gets or sets the stock quantity
         /// </summary>
-        public int StockQuantity { get; set; }
+        public decimal StockQuantity { get; set; }
 
         /// <summary>
         /// Gets or sets the reserved quantity (ordered but not shipped yet)
         /// </summary>
-        public int ReservedQuantity { get; set; }
+        public decimal ReservedQuantity { get; set; }
     }
 }
